Get the migrations going.
Ignore the storyboard changes. They are just an Xcode version thing.

The problem was that `RoomsViewController` was the root VC, and iOS was
initialising it then calling `applicationDidFInishingLaunching`. I’ve
changed that VC to initialise Realm only when it needs to in
`viewDidLoad`. Also I’ve put a todo that when you get the realm
notification you may need to regrab Rooms from Realm.